Move model to separate C library file
